Move Publish command to its own namespace
