+++ modules/cli/src/main/scala/scala/cli/ScalaCliCommands.scala
@@ -8,6 +8,7 @@ import java.nio.file.InvalidPathException
 import scala.cli.commands._
 import scala.cli.commands.github.{SecretCreate, SecretList}
 import scala.cli.commands.pgp.{PgpCommands, PgpCommandsSubst}
+import scala.cli.commands.publish.Publish
 
 class ScalaCliCommands(
   val progName: String,
